+++ src/main/scala/io/picnicml/doddlemodel/metrics/ClassificationMetrics.scala
@@ -9,8 +9,8 @@ object ClassificationMetrics {
 
     override lazy val higherValueIsBetter: Boolean = true
 
-    override def calculateValueSafe(y: Target, yPred: Target): Double =
-      (y :== yPred).activeSize / y.length.toDouble
+    override def calculateValueSafe(y: Target, yPred: Target): Float =
+      (y :== yPred).activeSize / y.length.toFloat
 
     override def toString: String = "accuracy"
   }
@@ -25,12 +25,12 @@ object ClassificationMetrics {
       require(numberOfTargetClasses(y) == 2, "Precision is defined for a binary classification task")
     }
 
-    override def calculateValueSafe(y: Target, yPred: Target): Double = {
-      val yPredPositive = yPred :== 1.0
-      val yPositive = y :== 1.0
+    override def calculateValueSafe(y: Target, yPred: Target): Float = {
+      val yPredPositive = yPred :== 1.0f
+      val yPositive = y :== 1.0f
 
-      val numTp = (yPredPositive &:& yPositive).activeSize.toDouble
-      numTp / (numTp + (yPredPositive &:& !yPositive).activeSize.toDouble)
+      val numTp = (yPredPositive &:& yPositive).activeSize.toFloat
+      numTp / (numTp + (yPredPositive &:& !yPositive).activeSize.toFloat)
     }
 
     override def toString: String = "precision"
@@ -46,12 +46,12 @@ object ClassificationMetrics {
       require(numberOfTargetClasses(y) == 2, "Recall is defined for a binary classification task")
     }
 
-    override def calculateValueSafe(y: Target, yPred: Target): Double = {
-      val yPredPositive = yPred :== 1.0
-      val yPositive = y :== 1.0
+    override def calculateValueSafe(y: Target, yPred: Target): Float = {
+      val yPredPositive = yPred :== 1.0f
+      val yPositive = y :== 1.0f
 
-      val numTp = (yPredPositive &:& yPositive).activeSize.toDouble
-      numTp / (numTp + (!yPredPositive &:& yPositive).activeSize.toDouble)
+      val numTp = (yPredPositive &:& yPositive).activeSize.toFloat
+      numTp / (numTp + (!yPredPositive &:& yPositive).activeSize.toFloat)
     }
 
     override def toString: String = "recall"
@@ -67,11 +67,11 @@ object ClassificationMetrics {
       require(numberOfTargetClasses(y) == 2, "F1 score is defined for a binary classification task")
     }
 
-    override def calculateValueSafe(y: Target, yPred: Target): Double = {
+    override def calculateValueSafe(y: Target, yPred: Target): Float = {
       val prec = precision(y, yPred)
       val rec = recall(y, yPred)
 
-      2.0 * (prec * rec) / (prec + rec)
+      2.0f * (prec * rec) / (prec + rec)
     }
 
     override def toString: String = "F1 score"
@@ -82,7 +82,7 @@ object ClassificationMetrics {
 
     override lazy val higherValueIsBetter: Boolean = false
 
-    override def calculateValueSafe(y: Target, yPred: Target): Double = 1.0 - accuracy(y, yPred)
+    override def calculateValueSafe(y: Target, yPred: Target): Float = 1.0f - accuracy(y, yPred)
 
     override def toString: String = "Hamming loss"
   }
