+++ src/test/scala/io/picnicml/doddlemodel/impute/MostFrequentValueImputerTest.scala
@@ -10,47 +10,47 @@ class MostFrequentValueImputerTest extends FlatSpec with Matchers with TestingUt
 
   "Most frequent value imputer" should "impute the categorical features" in {
     val xMissing = DenseMatrix(
-      List(Double.NaN, 1.0, 2.0),
-      List(3.0, Double.NaN, 5.0),
-      List(6.0, 7.0, 8.0),
-      List(6.0, 7.0, 2.0)
+      List(Float.NaN, 1.0f, 2.0f),
+      List(3.0f, Float.NaN, 5.0f),
+      List(6.0f, 7.0f, 8.0f),
+      List(6.0f, 7.0f, 2.0f)
     )
 
     val xImputedExpected = DenseMatrix(
-      List(6.0, 1.0, 2.0),
-      List(3.0, Double.NaN, 5.0),
-      List(6.0, 7.0, 8.0),
-      List(6.0, 7.0, 2.0)
+      List(6.0f, 1.0f, 2.0f),
+      List(3.0f, Float.NaN, 5.0f),
+      List(6.0f, 7.0f, 8.0f),
+      List(6.0f, 7.0f, 2.0f)
     )
 
     val featureIndex = FeatureIndex.apply(List(CategoricalFeature, NumericalFeature, CategoricalFeature))
     val imputer = MostFrequentValueImputer(featureIndex)
     val fittedImputer = ev.fit(imputer, xMissing)
 
-    breezeEqual(fittedImputer.mostFrequent.value, DenseVector(6.0, 2.0)) shouldBe true
+    breezeEqual(fittedImputer.mostFrequent.value, DenseVector(6.0f, 2.0f)) shouldBe true
     breezeEqual(ev.transform(fittedImputer, xMissing), xImputedExpected) shouldBe true
   }
 
   it should "impute a subset of categorical features" in {
     val xMissing = DenseMatrix(
-      List(Double.NaN, 1.0, 2.0),
-      List(3.0, Double.NaN, 5.0),
-      List(6.0, 7.0, 8.0),
-      List(6.0, 7.0, 2.0)
+      List(Float.NaN, 1.0f, 2.0f),
+      List(3.0f, Float.NaN, 5.0f),
+      List(6.0f, 7.0f, 8.0f),
+      List(6.0f, 7.0f, 2.0f)
     )
 
     val xImputedExpected = DenseMatrix(
-      List(Double.NaN, 1.0, 2.0),
-      List(3.0, 7.0, 5.0),
-      List(6.0, 7.0, 8.0),
-      List(6.0, 7.0, 2.0)
+      List(Float.NaN, 1.0f, 2.0f),
+      List(3.0f, 7.0f, 5.0f),
+      List(6.0f, 7.0f, 8.0f),
+      List(6.0f, 7.0f, 2.0f)
     )
 
     val featureIndex = FeatureIndex.categorical(List(1, 2))
     val imputer = MostFrequentValueImputer(featureIndex)
     val fittedImputer = ev.fit(imputer, xMissing)
 
-    breezeEqual(fittedImputer.mostFrequent.value, DenseVector(7.0, 2.0)) shouldBe true
+    breezeEqual(fittedImputer.mostFrequent.value, DenseVector(7.0f, 2.0f)) shouldBe true
     breezeEqual(ev.transform(fittedImputer, xMissing), xImputedExpected) shouldBe true
   }
 }
